backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.9
itsdangerous==2.2.0
cryptography==43.0.1
sqlmodel==0.0.21
alembic==1.13.2
python-dotenv==1.1.1
boto3==1.35.23

